% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gist_to_carbon.R
\docType{data}
\name{carbon_fonts}
\alias{carbon_fonts}
\title{Carbon fonts}
\format{
An object of class \code{character} of length 13.
}
\usage{
carbon_fonts
}
\description{
A list of available fonts that carbon supports.
}
\details{
Sourced from: <https://github.com/carbon-app/carbon/blob/main/lib/constants.js>
}
\examples{
{
carbon_fonts
}
}
\keyword{datasets}
