% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gist_to_carbon.R
\docType{data}
\name{carbon_themes}
\alias{carbon_themes}
\title{Carbon themes}
\format{
An object of class \code{character} of length 29.
}
\usage{
carbon_themes
}
\description{
A list of available themes supported by Carbon.
}
\details{
Sourced from <https://github.com/carbon-app/carbon/blob/main/lib/constants.js>
}
\examples{
{
carbon_themes
}
}
\keyword{datasets}
