% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gist_to_carbon.R
\name{gist_append_img}
\alias{gist_append_img}
\title{Add imgur-hosted URL to existing gist as a "comment" to bottom of script.}
\usage{
gist_append_img(imgur_url, gist_id = NULL)
}
\arguments{
\item{imgur_url}{Existing URL from imgur, typically as created with gistillery::gist_to_carbon()}

\item{gist_id}{Unique ID for an existing Github Gist - this is where the comment will be added.}
}
\value{
Adds a commented line to bottom of existing Gist code
}
\description{
Add imgur-hosted URL to existing gist as a "comment" to bottom of script.
}
