% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gist_to_carbon.R
\name{gist_to_carbon}
\alias{gist_to_carbon}
\title{Take an existing gist, send to carbon, and save the image locally}
\usage{
gist_to_carbon(
  gist_id,
  file = "code.png",
  bg = getOption("gistillery.bg", default = "#4A90E2"),
  theme = getOption("gistillery.theme", default = "night-owl"),
  font = getOption("gistillery.font", default = "Hack"),
  lang = "auto",
  imgur = TRUE,
  drop_shadow = TRUE,
  width = 680,
  width_auto_adjust = TRUE
)
}
\arguments{
\item{gist_id}{the unique id for your existing gist, can be piped from gist_upload() or passed manually.}

\item{file}{the name of the file for printing, eg code.png}

\item{bg}{A valid hex code for color, ie #D3D3D3}

\item{theme}{A valid theme, such as "cobalt", "nord", "night-owl", "monokai" - for all available themes, see `carbon_themes`}

\item{font}{A valid font ID such as "IBM+Plex+Mono", "Hack", "Fira+Code" - for all available fonts, see `carbon_fonts`}

\item{lang}{A language for syntax highlighting, ie one of "python", "r", "yaml", "markdown", "text", "auto"}

\item{imgur}{A logical, should the image also be uploaded to imgur.}

\item{drop_shadow}{Logical indicating whether to include drop shadow for the screenshot.}

\item{width}{a number, indicating the width in pixels for screenshot}

\item{width_auto_adjust}{Logical indicating whether to auto adjust the width for better code-printing}
}
\value{
Saves an image to disk and optionally returns the uploaded imgur URL
}
\description{
Take an existing gist, send to carbon, and save the image locally
}
