% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git_auth.R
\name{git_auth}
\alias{git_auth}
\title{Authorize with GitHub.}
\usage{
git_auth(app = gistr_app, reauth = FALSE)
}
\arguments{
\item{app}{An [httr::oauth_app()] for GitHub. The default uses an
application `gistr_oauth` created by Scott Chamberlain.}

\item{reauth}{(logical) Force re-authorization?}
}
\value{
a character string - used downstream for auth in various functions
}
\description{
This function is run automatically to allow gistillery to access your GitHub
account. It is adapted from gistr::gist_auth
}
\details{
There are two ways to authorise gistillery to work with your GitHub account:

- Generate a personal access token with the gist scope selected, and set it
as the `GITHUB_PAT` environment variable per session using `Sys.setenv`
or across sessions by adding it to your `.Renviron` file or similar.
See
https://help.github.com/articles/creating-an-access-token-for-command-line-use
for help
- Interactively login into your GitHub account and authorise with OAuth.

Using `GITHUB_PAT` is recommended.
}
