% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reprex_shot.R
\name{reprex_shot}
\alias{reprex_shot}
\title{Execute a reprex::reprex() and take a local screenshot of the reprex output}
\usage{
reprex_shot(filename = NULL, ..., open_file = TRUE, imgur = FALSE)
}
\arguments{
\item{filename}{a filename, ending in .png}

\item{...}{additional arguments passed to reprex::reprex()}

\item{open_file}{A logical, should the file be opened once saved}

\item{imgur}{A logical, should the image be uploaded to imgur also}
}
\value{
a screenshot of the reprex on disk
}
\description{
reprex_shot() will first take a reprex and then capture the
HTML output into an on-disk image, optionally uploading the image to Imgur.
}
