% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gist_comment.R
\name{gist_comment}
\alias{gist_comment}
\title{Add a comment to an existing gist}
\usage{
gist_comment(
  gist_id,
  body = "![](pbs.twimg.com/media/FBGfjADUYAUxiPz?format=png)",
  ...
)
}
\arguments{
\item{gist_id}{Character string indicating the gist ID}

\item{body}{The comment text}

\item{...}{Additional arguments passed to `gistr:::gist_post()`}
}
\value{
A string of the comment text
}
\description{
Add a comment to an existing gist
}
