% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gist_comment.R
\name{gist_comment}
\alias{gist_comment}
\title{Add a comment to an existing gist}
\usage{
gist_comment(
  gist_id = "73b34a7d484360036fee122c5a745f44",
  comment = "![](https://pbs.twimg.com/media/FBGfjADUYAUxiPz?format=png)"
)
}
\arguments{
\item{gist_id}{Character string indicating the gist ID}

\item{comment}{The comment text}
}
\value{
A string of the comment text
}
\description{
Add a comment to an existing gist
}
