% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gist_upload.R
\name{gist_upload}
\alias{gist_upload}
\title{Take local code and upload to a named gist}
\usage{
gist_upload(
  content = NULL,
  gist_name = NULL,
  description = "",
  public = TRUE,
  browse = TRUE
)
}
\arguments{
\item{content}{the code, either the currently highlighted file or manually indicated code}

\item{gist_name}{a valid filename ie my-code.R}

\item{description}{a brief description of the gist or it's purpose}

\item{public}{a logical, defaults to TRUE, indicates whether to make gist public or not}

\item{browse}{a logical, defaults to TRUE, indicates whether to open the new gist in browser or not}
}
\value{
gist id and the gist URL to clipboard, can be piped directly into gist_to_carbon
}
\description{
Take local code and upload to a named gist
}
