% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reprex_shot.R
\name{reprex_shot}
\alias{reprex_shot}
\title{Take a local screenshot of reprex output}
\usage{
reprex_shot(code = NULL, filename = NULL, open_file = TRUE, imgur = FALSE)
}
\arguments{
\item{code}{Optional code from a reprex, mainly here so that you can pipe reprex directly into this function.}

\item{filename}{a filename, ending in .png}

\item{open_file}{A logical, should the file be opened once saved}

\item{imgur}{A logical, should the image be uploaded to imgur also}
}
\value{
a screenshot of the reprex on disk
}
\description{
Take a local screenshot of reprex output
}
